Workaround: Use etds as previous stops, not next stops
Previously, this code was using etds incorrectly, computing
from the station instead of to the station.

Note that this commit should be fixed eventually because
it prevents trains headed to route endpoints from being processed
(because the candidateRoutes will have trains where the station
is the first train, so index - 1 is undefined).